+++ bittytax/conv/parsers/__init__.py
@@ -1,3 +1,4 @@
+from . import accointing
 from . import binance
 from . import barclays
 from . import bitfinex